Split out signIn impl
