+++ src/server/implementation/utils.ts
@@ -1,8 +1,4 @@
-import {
-  alphabet,
-  generateRandomString,
-  sha256 as rawSha256,
-} from "oslo/crypto";
+import { sha256 as rawSha256 } from "oslo/crypto";
 import { encodeHex } from "oslo/encoding";
 
 export const TOKEN_SUB_CLAIM_DIVIDER = "|";
@@ -14,4 +10,12 @@ export function stringToNumber(value: string | undefined) {
 
 export async function sha256(input: string) {
   return encodeHex(await rawSha256(new TextEncoder().encode(input)));
+}
+
+export function logError(error: unknown) {
+  console.error(
+    error instanceof Error
+      ? error.message + "\n" + error.stack?.replace("\\n", "\n")
+      : error,
+  );
 }
